remove redux from settings
