+++ client/src/Hooks/monitorHooks.js
@@ -4,6 +4,7 @@ import { createToast } from "../Utils/toastUtils";
 import { useTheme } from "@emotion/react";
 import { useMonitorUtils } from "./useMonitorUtils";
 import { useNavigate } from "react-router-dom";
+import { useTranslation } from "react-i18next";
 
 const useFetchMonitorsWithSummary = ({ teamId, types, monitorUpdateTrigger }) => {
 	const [isLoading, setIsLoading] = useState(false);
@@ -383,6 +384,58 @@ const usePauseMonitor = () => {
 	return [pauseMonitor, isLoading, error];
 };
 
+const useAddDemoMonitors = () => {
+	const [isLoading, setIsLoading] = useState(false);
+	const { t } = useTranslation();
+	const addDemoMonitors = async () => {
+		try {
+			setIsLoading(true);
+			await networkService.addDemoMonitors();
+			createToast({ body: t("settingsDemoMonitorsAdded") });
+		} catch (error) {
+			createToast({ body: t("settingsFailedToAddDemoMonitors") });
+		} finally {
+			setIsLoading(false);
+		}
+	};
+	return [addDemoMonitors, isLoading];
+};
+
+const useDeleteAllMonitors = () => {
+	const [isLoading, setIsLoading] = useState(false);
+	const { t } = useTranslation();
+	const deleteAllMonitors = async () => {
+		try {
+			setIsLoading(true);
+			await networkService.deleteAllMonitors();
+			createToast({ body: t("settingsMonitorsDeleted") });
+		} catch (error) {
+			createToast({ body: t("settingsFailedToDeleteMonitors") });
+		} finally {
+			setIsLoading(false);
+		}
+	};
+	return [deleteAllMonitors, isLoading];
+};
+
+const UseDeleteMonitorStats = () => {
+	const { t } = useTranslation();
+	const [isLoading, setIsLoading] = useState(false);
+	const deleteMonitorStats = async ({ teamId }) => {
+		setIsLoading(true);
+		try {
+			const res = await networkService.deleteChecksByTeamId({ teamId });
+			createToast({ body: t("settingsStatsCleared") });
+		} catch (error) {
+			createToast({ body: t("settingsFailedToClearStats") });
+		} finally {
+			setIsLoading(false);
+		}
+	};
+
+	return [deleteMonitorStats, isLoading];
+};
+
 export {
 	useFetchMonitorsWithSummary,
 	useFetchMonitorsWithChecks,
@@ -395,4 +448,7 @@ export {
 	useDeleteMonitor,
 	useUpdateMonitor,
 	usePauseMonitor,
+	useAddDemoMonitors,
+	useDeleteAllMonitors,
+	UseDeleteMonitorStats,
 };
